crosscluster: move aggregate range stats collector into utils
This patch moves the aggregate range stats collector for LDR into the
replication utils package so that it can be used in PCR.

Epic: None

Release note: None
