+++ pkg/crosscluster/logical/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
         "metrics.go",
         "offline_initial_scan_processor.go",
         "purgatory.go",
-        "range_stats.go",
         "replication_statements.go",
         "savepoint.go",
         "sql_crud_writer.go",
@@ -128,7 +127,6 @@ go_test(
         "lww_row_processor_test.go",
         "main_test.go",
         "purgatory_test.go",
-        "range_stats_test.go",
         "replication_statements_test.go",
         "savepoint_test.go",
         "sql_row_reader_test.go",
